+++ pyorc/velocimetry/__init__.py
@@ -1,6 +1,5 @@
 """pyorc velocimetry methods."""
 
 from .ffpiv import get_ffpiv
-from .openpiv import get_openpiv, piv
 
-__all__ = ["get_ffpiv", "piv", "get_openpiv"]
+__all__ = ["get_ffpiv"]